convert all to int

+++ validate.py
@@ -249,7 +249,7 @@ def validate_multi(val_loader, model, args):
           .format(mean_p_c, mean_r_c, mean_f_c, p_o, r_o, f_o))
 
     mAP_score = mAP(torch.cat(targets).numpy(), torch.cat(preds).numpy())
-    wf1 = f1_score(torch.cat(targets).numpy(), torch.cat(preds).numpy(), average = "weighted")
+    wf1 = f1_score(torch.cat(targets).numpy().astype(int), torch.cat(preds).numpy().astype(int), average = "weighted")
     print("mAP score:", mAP_score)
     print("weightedF1 score:", wf1)
     #mlflow
